requirements.txt
#django-redis
beautifulsoup4==4.10.0
Django==3.2.7
django-ckeditor==6.1.0
django-comments-xtd==2.9.3
django_compressor==2.4.1
django_contrib_comments==2.1.0
django-filer==2.0.2
django-grappelli==2.15.1
-e git+https://github.com/maximst/django-hvad.git#egg=django-hvad
django-imagekit==4.0.2
django-mptt==0.13.3
django-registration-redux==2.9
django-simple-captcha==0.5.14
django-static-precompiler==2.0
django-taggit==1.5.1
django-voting==1.0
-e git+https://github.com/maximst/LinkExchange.git#egg=LinkExchange
-e git+https://github.com/maximst/LinkExchange.Django.git#egg=LinkExchange.Django
Pillow==8.3.2
psycopg2==2.9.1
python-facebook-api==0.11.2
python-social-auth==0.3.6
pytils==0.3
pytz
requests==2.26.0
social-auth-app-django==5.0.0
sorl-thumbnail==12.7.0
#-e git+https://github.com/maximst/supercaptcha.git#egg=supercaptcha
vkontakte==1.3.5
